# swagger/api/openapi.yaml
openapi: 3.0.0
info:
  title: Simple Inventory API
  description: This is a simple API
  contact:
    email: you@your-company.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: https://virtserver.swaggerhub.com/KZRT0123_1/showyourhands/1.0.0
    description: SwaggerHub API Auto Mocking
tags:
  - name: admins
    description: Secured Admin-only calls
  - name: developers
    description: Operations available to regular developers
paths:
  /login:
    get:
      tags:
        - front
      summary: shows login page
      description: shows login page
      operationId: getLoginPage
      responses:
        "200":
          description: login page updated
        "400":
          description: cannot load login page
      x-swagger-router-controller: Front
    post:
      tags:
        - query
      summary: login attempt
      description: check id and password
      operationId: loginRequest
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginProfile'
        description: login profile
        required: true
      responses:
        "201":
          description: item created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginToken'
        "400":
          description: server dead
        "401":
          description: login failed
      x-swagger-router-controller: Query
  /signup:
    get:
      tags:
        - front
      summary: shows signup page
      description: shows signup page
      operationId: getSignupPage
      responses:
        "200":
          description: Signup page updated
        "400":
          description: cannot load Signup page
      x-swagger-router-controller: Front
    post:
      tags:
        - query
      summary: signup attempt
      description: insert id pw in db
      operationId: signupRequest
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupProfile'
        description: signup profile
        required: true
      responses:
        "201":
          description: item created, return to login screen
        "400":
          description: server dead
        "401":
          description: signup failed
      x-swagger-router-controller: Query
  /game/start:
    get:
      tags:
        - query
      summary: fetch one word
      description: fetch one word, return count
      operationId: gameStartRequest
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientSideToken'
        description: client side token
        required: true
      responses:
        "200":
          description: word fetched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameWord'
        "400":
          description: game start failed
        "401":
          description: game ended
      x-swagger-router-controller: Query
  /game/next:
    get:
      tags:
        - query
      summary: fetch next word
      description: fetch one word, return count. if count > X return 401
      operationId: gameStartRequest
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientSideToken'
        description: client side token
        required: true
      responses:
        "200":
          description: word fetched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameWord'
        "400":
          description: game start failed
        "401":
          description: game ended
      x-swagger-router-controller: Query

components:
  schemas:
    SignupProfile:
      required:
        - email
        - id
        - nickname
        - password
      type: object
      properties:
        id:
          type: string
          format: VARCHAR(255)
          example: asap0123
        password:
          type: string
          format: VARCHAR(255)
          example: asap0123!
        nickname:
          type: string
          format: VARCHAR(20)
          example: asap
        email:
          type: string
          format: "VARCHAR(255), !do not check at the backend!"
          example: asap@asap.com
    LoginProfile:
      required:
        - id
        - password
      type: object
      properties:
        id:
          type: string
          format: VARCHAR(255)
          example: asap0123
        password:
          type: string
          format: VARCHAR(255)
          example: asap0123!
    LoginToken:
      required:
        - nickname
        - token
      type: object
      properties:
        nickname:
          type: string
          format: VARCHAR(20)
          example: asap
        token:
          type: string
          format: JWT
          example: arstarst
    ClientSideToken:
      required:
        - api
        - count
      type: object
      properties:
          api:
            type: string
            format: VARCHAR(128)
            example: arstarst
          count:
            type: integer
            format: INT
            example: 0

    GameWord:
      required:
        - word
        - count
      type: object
      properties:
        word:
          type: string
          format: VARCHAR(255)
          example: 인사
        count:
          type: integer
          format: INT
          example: 0